↪️ app route propagator
