+++ packages/react-shopify-app-route-propagator/CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-<!-- Unreleased changes should go to UNRELEASED.md -->
+## [Unreleased]
 
----
+### Changed
+
+- Breaking change - Update to use react hook @ `react` 16.8.1 or higher, also `app-bridge` at 1.0.0. or higher ([#616](https://github.com/Shopify/quilt/pull/616))
 
 ## 2.0.0 - 2019-01-09
