+++ packages/react-shopify-app-route-propagator/src/globals.ts
@@ -9,3 +9,5 @@ export function getTopWindow() {
 export function getOrigin() {
   return location.origin;
 }
+
+export const MODAL_IFRAME_NAME = 'app-modal-iframe';
