+++ packages/react-shopify-app-route-propagator/src/types.ts
@@ -1,0 +1,3 @@
+export type LocationOrHref =
+  | string
+  | {search: string; hash: string; pathname: string};
